+++ cypress/integration/blocks-listing.js
@@ -48,19 +48,13 @@ if (Cypress.env('API') !== 'guillotina') {
       cy.waitForResourceToLoad('my-page?fullobjects');
       cy.navigate('/my-page/edit');
 
-      cy.get(`.block.title [data-contents]`)
-        .clear()
-        .type('My title');
+      cy.get(`.block.title [data-contents]`).clear().type('My title');
 
       //add listing block
       cy.get('.block.text [contenteditable]').click();
       cy.get('button.block-add-button').click();
-      cy.get('.blocks-chooser .title')
-        .contains('Common')
-        .click();
-      cy.get('.blocks-chooser .common')
-        .contains('Listing')
-        .click();
+      cy.get('.blocks-chooser .title').contains('Common').click();
+      cy.get('.blocks-chooser .common').contains('Listing').click();
 
       //verify before save
       cy.get(`.block.listing .listing-body:first-of-type`).contains(
@@ -114,19 +108,13 @@ if (Cypress.env('API') !== 'guillotina') {
       cy.waitForResourceToLoad('my-page?fullobjects');
       cy.navigate('/my-page/edit');
 
-      cy.get(`.block.title [data-contents]`)
-        .clear()
-        .type('My title');
+      cy.get(`.block.title [data-contents]`).clear().type('My title');
 
       //add listing block
       cy.get('.block.text [contenteditable]').click();
       cy.get('button.block-add-button').click();
-      cy.get('.blocks-chooser .title')
-        .contains('Common')
-        .click();
-      cy.get('.blocks-chooser .common')
-        .contains('Listing')
-        .click();
+      cy.get('.blocks-chooser .title').contains('Common').click();
+      cy.get('.blocks-chooser .common').contains('Listing').click();
 
       //verify before save
       cy.get(`.block.listing .listing-body:first-of-type`).contains('Page One');
@@ -186,20 +174,14 @@ if (Cypress.env('API') !== 'guillotina') {
       //add listing block
       cy.get('.block.text [contenteditable]').click();
       cy.get('button.block-add-button').click();
-      cy.get('.blocks-chooser .title')
-        .contains('Common')
-        .click();
-      cy.get('.blocks-chooser .common')
-        .contains('Listing')
-        .click();
+      cy.get('.blocks-chooser .title').contains('Common').click();
+      cy.get('.blocks-chooser .common').contains('Listing').click();
 
       //********  add short-name criteria filter
       cy.get('.sidebar-container .tabs-wrapper .menu .item')
         .contains('Block')
         .click();
-      cy.get('.sidebar-listing-data .fields')
-        .contains('Add criteria')
-        .click();
+      cy.get('.sidebar-listing-data .fields').contains('Add criteria').click();
       cy.get(
         '.sidebar-listing-data .fields:first-of-type .field:first-of-type .react-select__menu .react-select__option',
       )
@@ -279,20 +261,14 @@ if (Cypress.env('API') !== 'guillotina') {
       //add listing block
       cy.get('.block.text [contenteditable]').click();
       cy.get('button.block-add-button').click();
-      cy.get('.blocks-chooser .title')
-        .contains('Common')
-        .click();
-      cy.get('.blocks-chooser .common')
-        .contains('Listing')
-        .click();
+      cy.get('.blocks-chooser .title').contains('Common').click();
+      cy.get('.blocks-chooser .common').contains('Listing').click();
 
       //********  add location criteria filter
       cy.get('.sidebar-container .tabs-wrapper .menu .item')
         .contains('Block')
         .click();
-      cy.get('.sidebar-listing-data .fields')
-        .contains('Add criteria')
-        .click();
+      cy.get('.sidebar-listing-data .fields').contains('Add criteria').click();
       cy.get(
         '.sidebar-listing-data .fields:first-of-type .field:first-of-type .react-select__menu .react-select__option',
       )
@@ -371,20 +347,14 @@ if (Cypress.env('API') !== 'guillotina') {
       //add listing block
       cy.get('.block.text [contenteditable]').click();
       cy.get('button.block-add-button').click();
-      cy.get('.blocks-chooser .title')
-        .contains('Common')
-        .click();
-      cy.get('.blocks-chooser .common')
-        .contains('Listing')
-        .click();
+      cy.get('.blocks-chooser .title').contains('Common').click();
+      cy.get('.blocks-chooser .common').contains('Listing').click();
 
       //********  add location criteria filter
       cy.get('.sidebar-container .tabs-wrapper .menu .item')
         .contains('Block')
         .click();
-      cy.get('.sidebar-listing-data .fields')
-        .contains('Add criteria')
-        .click();
+      cy.get('.sidebar-listing-data .fields').contains('Add criteria').click();
       cy.get(
         '.sidebar-listing-data .fields:first-of-type .field:first-of-type .react-select__menu .react-select__option',
       )
@@ -425,7 +395,87 @@ if (Cypress.env('API') !== 'guillotina') {
         .contains('Document outside Folder')
         .should('not.exist');
     });
-    it('Listing block - Test Criteria: Location Navigation path', () => {
+    it('Listing block - Test Criteria: Location relative particular to a issue', () => {
+      cy.createContent({
+        contentType: 'Document',
+        contentId: 'about-us',
+        contentTitle: 'about us',
+        path: '/',
+      });
+      cy.createContent({
+        contentType: 'Document',
+        contentId: 'contact',
+        contentTitle: 'contact',
+        path: '/',
+      });
+
+      cy.createContent({
+        contentType: 'Document',
+        contentId: 'news-item-one',
+        contentTitle: 'News Item One',
+        path: 'my-page',
+      });
+
+      cy.createContent({
+        contentType: 'Document',
+        contentId: 'news-item-two',
+        contentTitle: 'News Item Two',
+        path: 'my-page',
+      });
+
+      //add listing block
+      cy.get('.block.text [contenteditable]').click();
+      cy.get('button.block-add-button').click();
+      cy.get('.blocks-chooser .title').contains('Common').click();
+      cy.get('.blocks-chooser .common').contains('Listing').click();
+
+      //********  add location criteria filter
+      cy.get('.sidebar-container .tabs-wrapper .menu .item')
+        .contains('Block')
+        .click();
+      cy.get('.sidebar-listing-data .fields').contains('Add criteria').click();
+      cy.get(
+        '.sidebar-listing-data .fields:first-of-type .field:first-of-type .react-select__menu .react-select__option',
+      )
+        .contains('Location')
+        .click();
+      //location relative..
+      cy.get(
+        '.sidebar-listing-data .fields:first-of-type .main-fields-wrapper .field:last-of-type',
+      ).click();
+      cy.get(
+        '.sidebar-listing-data .fields:first-of-type .main-fields-wrapper .field:last-of-type .react-select__menu .react-select__option',
+      )
+        .contains('Relative path')
+        .click();
+
+      //insert relative path
+      cy.get('.sidebar-listing-data .fields:first-of-type > .field input')
+        .clear()
+        .type('.');
+
+      // verify if in list there's a page with name "Document within Folder"
+      cy.get(`.block.listing .listing-body:first-of-type`).contains(
+        'News Item One',
+      );
+      // verify if in list there isn't page with name "Document outside Folder"
+      cy.get(`.block.listing .listing-body`)
+        .contains('about us')
+        .should('not.exist');
+
+      // save;
+      cy.get('#toolbar-save').click();
+
+      //test location relative criteria after save
+      cy.get(`.block.listing .listing-body:first-of-type`).contains(
+        'News Item One',
+      );
+      cy.get(`.block.listing .listing-body`)
+        .contains('about us')
+        .should('not.exist');
+    });
+
+    it('Listing block - Test Criteria: Location Navigation', () => {
       /*not implemented because Navigation ui is not yet developed in Listing Block sidebar*/
     });
   });